fixed float

+++ inc/libft.h
@@ -87,7 +87,7 @@ void				ft_lstadd(t_list **alst, t_list *new);
 void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list				*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
 int					ft_isspace(int c);
-int					ft_pow(int a, int b);
+long long int 		ft_pow(int a, int b);
 int					ft_wrdcnt(char const *s, char c);
 int					ft_wordlen(const char *s, char c, int start);
 void				ft_lstrev(t_list **head);
